+++ lchflags.c
@@ -1,0 +1,55 @@
+/*
+ * Copyright 2022 Gaël PORTAY
+ *
+ * SPDX-License-Identifier: LGPL-2.1-or-later
+ */
+
+#ifdef __FreeBSD__
+#include <stdio.h>
+#include <errno.h>
+#include <limits.h>
+#include <fcntl.h>
+#include <dlfcn.h>
+
+#include <sys/stat.h>
+#include <unistd.h>
+
+#include "iamroot.h"
+
+__attribute__((visibility("hidden")))
+int next_lchflags(const char *path, unsigned long flags)
+{
+	int (*sym)(const char *, unsigned long);
+	int ret;
+
+	sym = dlsym(RTLD_NEXT, "lchflags");
+	if (!sym) {
+		__dlperror(__func__);
+		errno = ENOSYS;
+		return -1;
+	}
+
+	ret = sym(path, flags);
+	if (ret == -1)
+		__pathperror(path, __func__);
+
+	return ret;
+}
+
+int lchflags(const char *path, unsigned long flags)
+{
+	char buf[PATH_MAX];
+	ssize_t siz;
+
+	siz = path_resolution(AT_FDCWD, path, buf, sizeof(buf),
+			      AT_SYMLINK_NOFOLLOW);
+	if (siz == -1) {
+		__pathperror(path, __func__);
+		return -1;
+	}
+
+	__debug("%s(path: '%s', ...)\n", __func__, path);
+
+	return next_lchflags(path, flags);
+}
+#endif