+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   `__lxstat64()`, `__xstat64()`, `fopen64()`, `freopen64()`, `fstat64()`,
   `fstatat64()`, `lstat64()`, `open64()`, `openat64()` and `stat64()` for
   [musl] world
+- Intercept the [FreeBSD]'s functions `chflags()`, `chflagsat` and `lchflags()`
 
 ### Changed